+++ .circleci/config.yml
@@ -1,20 +1,44 @@
 ---
 version: 2
-jobs:
-  build:
+
+references:
+  bundle_install: &bundle_install
+    run:
+      name: Bundle
+      command: bundle install --path vendor/bundle --jobs=4 --retry=3
+  cache_bundle_for_lockfile: &cache_bundle_for_lockfile
+    save_cache:
+      key: v3-ruby2.4.1-Gemfile-{{ checksum "Gemfile.lock" }}
+      paths:
+        - vendor/bundle
+  cache_bundle: &cache_bundle
+    save_cache:
+      key: v3-ruby2.4.1-Gemfile
+      paths:
+        - vendor/bundle
+  restore_bundle: &restore_bundle
+    restore_cache:
+      keys:
+        - v3-ruby2.4.1-Gemfile-{{ checksum "Gemfile.lock" }}
+        - v3-ruby2.4.1-Gemfile
+  docker_build_image: &docker_build_image
     docker:
-      - image: circleci/golang:1.9
+      - image: gocardless/draupnir-circleci:latest
     working_directory: /go/src/github.com/gocardless/draupnir
+
+jobs:
+  unit:
+    <<: *docker_build_image
     environment:
       TEST_RESULTS: /tmp/test-results
+      GOPATH: /go
     steps:
       - checkout
       - run: go get github.com/golang/dep/cmd/dep
       - run: dep ensure -v && dep prune && test -z "$(git status --porcelain -- Gopkg.toml Gopkg.lock vendor)"
       - run: mkdir -p $TEST_RESULTS
       - run: make test
-      - run: make build-production
-      - run: make client
+
   rubocop:
     docker:
       - image: circleci/ruby:2.4.2
@@ -23,9 +47,44 @@ jobs:
       - checkout
       - run: bundle install
       - run: bundle exec rubocop
+
+  build-linux:
+    <<: *docker_build_image
+    environment:
+      GOPATH: /go
+    steps:
+      - checkout
+      - run:
+          name: Build linux binary
+          command: make build-production
+      - persist_to_workspace:
+          root: /go/src/github.com/gocardless/draupnir
+          paths: ['*.linux_amd64']
+
+  integration:
+    machine: true
+    steps:
+      - checkout
+      - attach_workspace:
+          at: workspace
+      - restore_cache:
+          key: dependency-cache-{{ checksum "Gemfile.lock" }}
+      - run:
+          name: Install Ruby bundle
+          command: bundle check --path=vendor/bundle || bundle install --path=vendor/bundle --jobs=4 --retry=3
+      - run:
+          name: Pull draupnir base image
+          command: docker pull gocardless/draupnir-base
+      - run:
+          name: Run integration tests
+          command: cp workspace/*.linux_amd64 ./ && make test-integration
+
 workflows:
   version: 2
-  tests:
+  build-integration:
     jobs:
-      - build
+      - unit
       - rubocop
+      - build-linux
+      - integration:
+          requires: [build-linux]
