--- .kitchen.yml
@@ -1,35 +1,0 @@
----
-driver:
-  name: vagrant
-  customize:
-    createhd:
-      filename: "tmp/%{instance_name}.vdi"
-      size: 512000
-    storageattach:
-      storagectl: SATA Controller
-      port: 1
-      device: 0
-      type: hdd
-      medium: "tmp/%{instance_name}.vdi"
-  synced_folders:
-    - [".", "/vagrant"]
-  network:
-    - ["private_network", {ip: "192.168.2.3"}]
-
-provisioner:
-  name: chef_zero_encrypt_databags
-  data_bags_path: chef/data_bags
-  encrypted_data_bag_secret_key_path: chef/data_bags/key
-
-platforms:
-  - name: ubuntu-16.04
-
-suites:
-  - name: default
-    run_list:
-      - recipe[draupnir::default]
-    attributes:
-      draupnir:
-        port: 8443
-        install_from_local_package: true
-        local_package_path: /vagrant/draupnir_1.0.0_amd64.deb
