--- Berksfile
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-source "https://supermarket.chef.io"
-
-cookbook "draupnir", path: "tmp/cookbooks/draupnir/berks-cookbooks/draupnir"
-cookbook "sudo", path: "tmp/cookbooks/draupnir/berks-cookbooks/sudo"
-cookbook "iptables-ng", path: "tmp/cookbooks/draupnir/berks-cookbooks/iptables-ng"
