+++ Gemfile
@@ -2,12 +2,11 @@
 
 source 'https://rubygems.org'
 
-gem 'berkshelf'
+gem 'activesupport'
+gem 'docker-api'
 gem 'fpm'
-gem 'kitchen-encrypt-databags'
-gem 'kitchen-vagrant'
+gem 'pry'
 gem 'rest-client'
 gem 'rspec'
 gem 'ruboconfig', git: 'https://github.com/gocardless/ruboconfig', branch: 'v1.0.1'
 gem 'rubocop'
-gem 'test-kitchen'
