--- chef/data_bags/key
@@ -1,1 +1,0 @@
-0
