--- fixtures/cert.pem
@@ -1,38 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIGqjCCBJKgAwIBAgIJAPJqpylVkLylMA0GCSqGSIb3DQEBBQUAMIGUMQswCQYD
-VQQGEwJHQjEPMA0GA1UECBMGTG9uZG9uMQ8wDQYDVQQHEwZMb25kb24xFzAVBgNV
-BAoTDkdvQ2FyZGxlc3MgTHRkMRMwEQYDVQQLEwpHb0NhcmRsZXNzMREwDwYDVQQD
-EwhkcmF1cG5pcjEiMCAGCSqGSIb3DQEJARYTaG1hY0Bnb2NhcmRsZXNzLmNvbTAe
-Fw0xNzA5MDcxNzEyNDRaFw0xODA5MDcxNzEyNDRaMIGUMQswCQYDVQQGEwJHQjEP
-MA0GA1UECBMGTG9uZG9uMQ8wDQYDVQQHEwZMb25kb24xFzAVBgNVBAoTDkdvQ2Fy
-ZGxlc3MgTHRkMRMwEQYDVQQLEwpHb0NhcmRsZXNzMREwDwYDVQQDEwhkcmF1cG5p
-cjEiMCAGCSqGSIb3DQEJARYTaG1hY0Bnb2NhcmRsZXNzLmNvbTCCAiIwDQYJKoZI
-hvcNAQEBBQADggIPADCCAgoCggIBANT+VlXFiBhGrbHAH1TPW9wT19D71bIb9uIJ
-zdtAddefe3+BgPSwxRRxuvjEy8bkruqUOf4Rx6EHXEMqq21rgEICZizasbW4t8FT
-iRw0dfzUxAjPdyB3pRDUAExhHs1nNfemJStTPKItYZqmYz8WZBsN9zpUSon0/ujU
-LkQv3wyNmRvjKeHvEfeweBVmL818W5ssiFqtHYGAIDsOXZnWaSgJISDowtRop+9o
-nn0ojaXM1HiCo2mP2dF/3h54hNpjX0BcngGY61NEPcWdsmzXfcL7XPcmo7x/rP4O
-3gwdijaRd3h91qMRu7TTn/1OoRwJsw2/7akKhL7r/v6Bya6VxD3cbqbkwVCRCKXC
-fn6RjGYvNs9cMlnT1tAUzXeXOKQiY6hRIUV/HWC8xjyvhHxgFaxSLiWSAIEgDo+J
-/D9lFbC+rBXQ+QUBlpaLy3keyk4q9927qrcz5K7tIt5IKf5T24o7gU4IaaLFNCEC
-bmmFtpOP6xNjVTHY5JnqkjG4xPKsV0dv23F3OXn6hcU7LKGUxa/tkM113cuvDwS9
-EtsmQfWQ0tvuEHNNSXKnK7IhIiR5/nMwXjYTP0hRYweqG1BjKYp0SJHoPe2UbHZB
-LUKI5ZipXyxqQlyWIZo3YhnD9WLkaVRXiVSL5Z2tcRpAGO/3UiUjlrMEIKUbVama
-TCof83+TAgMBAAGjgfwwgfkwHQYDVR0OBBYEFOsHHPRxF3app/3j7be4JGdrd7T1
-MIHJBgNVHSMEgcEwgb6AFOsHHPRxF3app/3j7be4JGdrd7T1oYGapIGXMIGUMQsw
-CQYDVQQGEwJHQjEPMA0GA1UECBMGTG9uZG9uMQ8wDQYDVQQHEwZMb25kb24xFzAV
-BgNVBAoTDkdvQ2FyZGxlc3MgTHRkMRMwEQYDVQQLEwpHb0NhcmRsZXNzMREwDwYD
-VQQDEwhkcmF1cG5pcjEiMCAGCSqGSIb3DQEJARYTaG1hY0Bnb2NhcmRsZXNzLmNv
-bYIJAPJqpylVkLylMAwGA1UdEwQFMAMBAf8wDQYJKoZIhvcNAQEFBQADggIBAJGM
-s7YCg+YdqQsX5MQoOstvPSXoxEwFjU+/L/3eoF8Eh1UyrANUAAFcmNUb3oT1SnNO
-Qo4YUuoVpfz8WM4mXod5dedRs9Y3qK05U1Dv4vFyJXsAG9RHLD5+bdQdRHVJgrtC
-mvsEKJt4Q0RxMsUmRfQpKQaen+Sr2ibfQdHF+i3n7P84KQ3sByEXZxynadYX7bao
-a3ZSAHrRX+SPd2QZbppDEgL4u2LFnsLCLBQ/3wXhLuct64usBjjP/N/PRtJuRsgB
-3twhsjklJ76SDuk2nG2E7r2p+Wy0l9ZZZ8tLjIM2yu/etwE+0Ay/9baKtZs9KauD
-NhrMh4zPeM2knRed6j9oFSsWjFwp4CRXbuwyT3ZutOoX/ge6g3F46i6+hvpcWfau
-JTCv+n1TaEIE2Ldc01kR0U8iowlABD92dK9ZSMhB/pkfENFqBiHqeyxXhMWG0/FG
-vNu4x39d14d2sWR95wyHlIACwNwpZNvAnX92Kclv0ff2ZKCRHZbx30ozrDabf8dC
-Ks+7SQF09fONh9dm3ehEi1wyH/FGq0HsDF7qvu1VHQ44cq2ghLEM6FKy12ie13NG
-ApWRiiVc0EbqDJ3VGX+8s9ji2XmdftFNtK0AGKl6AHlW8N6VaOJlj+YHI7f9+XGr
-UjPHAF+KkpBgZaJpArvfeFPYJJjYkyZz7tA61g3x
------END CERTIFICATE-----
