+++ spec/draupnir/access_token_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require "spec_helper"
+
+RSpec.describe "/access_tokens" do
+  describe "GET /authenticate" do
+    # TODO!
+    # Ideally we want to test the full oauth flow here.
+    # Unsure how we automate the browser bit though..
+  end
+end
