+++ spec/fixtures/key.pem
@@ -1,0 +1,51 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIIJKgIBAAKCAgEA1P5WVcWIGEatscAfVM9b3BPX0PvVshv24gnN20B11597f4GA
+9LDFFHG6+MTLxuSu6pQ5/hHHoQdcQyqrbWuAQgJmLNqxtbi3wVOJHDR1/NTECM93
+IHelENQATGEezWc196YlK1M8oi1hmqZjPxZkGw33OlRKifT+6NQuRC/fDI2ZG+Mp
+4e8R97B4FWYvzXxbmyyIWq0dgYAgOw5dmdZpKAkhIOjC1Gin72iefSiNpczUeIKj
+aY/Z0X/eHniE2mNfQFyeAZjrU0Q9xZ2ybNd9wvtc9yajvH+s/g7eDB2KNpF3eH3W
+oxG7tNOf/U6hHAmzDb/tqQqEvuv+/oHJrpXEPdxupuTBUJEIpcJ+fpGMZi82z1wy
+WdPW0BTNd5c4pCJjqFEhRX8dYLzGPK+EfGAVrFIuJZIAgSAOj4n8P2UVsL6sFdD5
+BQGWlovLeR7KTir33buqtzPkru0i3kgp/lPbijuBTghposU0IQJuaYW2k4/rE2NV
+MdjkmeqSMbjE8qxXR2/bcXc5efqFxTssoZTFr+2QzXXdy68PBL0S2yZB9ZDS2+4Q
+c01JcqcrsiEiJHn+czBeNhM/SFFjB6obUGMpinRIkeg97ZRsdkEtQojlmKlfLGpC
+XJYhmjdiGcP1YuRpVFeJVIvlna1xGkAY7/dSJSOWswQgpRtVqZpMKh/zf5MCAwEA
+AQKCAgEA0b8tEgYWAaIIhpSlxgT1dwJrczest1rmgq4Y9KLD6XT6TxQfA9fYHEvr
+TfOuoMyum/8gJN0pc6Al8Vbhvj0w7tAiAmXSvxuKiW1UBlx7HGICMpLrwyNPgATl
+zlGN3cSX1qrcy4+KlEKQbZURAx1vs1zLa7t6M+T56hqY/TkPM/N1JYtMwVuGnB5C
+AwDAbmDabTMri/ASsGaWVtHMxyPoJi7nWtsckcQGLLMf8CyxYOcXDR3JArlQTQjY
+mbUhT1awhFcwmm4Wv8bh9etXu+txw0N5x0Fb2NL6bq4BaHjmlNonhpaVebyK1oXY
+dcCq00uR8iZhSig7lTAdyk/RZoXB20TQhVewHbTg5Ckn/vWt7194XvF2LHhsXHGr
+stHPe/ej3KaHmdbemi92A2TDzcevo+2lfGAdYpLp0R1pKMLAkaOphVWesRwLK+9s
+bZbBRiGbqX+Qs5KOGshgbXUhAh2wBcfhRkgJ3OKQ9tWKIrooa/EurStKhc+v9PsP
+cpKPw3XrMuOgVLQulEsYce4eNsd0q1FQx8LYCWAzhLA8GOQIu12qRKZAmXFZoLtl
+HgSe6Zy7szQy2Yh7L1SN/W4oT3Szsq01zPsimLcD/7YgE50shrt5gbbHWw3xPJGV
+nWgmgYMpZ8Fi+0jPBs3FnXIyt/rjF9Yu9ecFcWsh/FgpphrRAGECggEBAPax1Dz1
+hbB2WOJbMkcXv0Ynso0yGRE3vkJGAfox/AoYqBFvFNA2KkFWWV/zYjal64UM3+lA
+F5HIxkFHQyCZEzFLFjrGNNB6ZQsI1Gubaiov1x3aSfKsWy4p5xLHHVujWlfY9cGH
+IqsO/Tcxt9ITbsH43RGWUH83gwQzgEJLD779cSiqIkRBbqJhq+3vQfayz+2h0Fxn
+/CG1lZoWTEUBclJTbZOe6u0FVE5kMysffdYypv08d7Pxp1NPp7OWwkU6h3HJ1N+x
+VSH+eTQD+HO5vzsQm9axWNyCgF5aqJLtIB4+Aq+HioQO8zY01byOE3ocH4ZRkvP2
+S5kUV0UbRmmI9BkCggEBAN0HE/o9vbQ4zvU1SEkqu7CpgkzTqPxXQPxcZT2wsTCg
+B29Rl7wsgZASIPENcl3Ofod8Wrc9IRFih05EYyOdIe0zZM2k8F/WEif2GkxziLgI
+oV3e8rpMq7dwYu3FOR0mA5L+1UjoPY5KEl7PBeOKv5nOVBPJpjr9snnbO1X0xfjJ
+w9Bi7lDC57WLfVap+BK+QVdt/R/vQMR5cQaZGgeWVm1ilZTZ++CXUJmqmutD1IDy
+oade1tFFNFbBxduZ4aSKnkDLIKPsvWjqLnu+mm8qjHQ3XO/4LAuCCgqWdb8F496F
+VajSOfBQpmaW9Ncgi9k7sP4Z0SSEV+dsUHW9yr6dZosCggEAKWBKKK9FbQNshQqm
+tCXXfvzJhcURT0H6N0NmsYAQteDMz5/B7eonzC2fb+hyYJgTmI1AQO7vBRSXq/s3
+1w/VmD7vy4vS7+SXbfJdLm8yr0C249ro7DrcUDHW/99FnUkYrgNFtN/lgKtS7R75
+pdJ8kSwcaAa+GZhF8T5yFKxiR7ZadGeUt74BwFqBa7blaWs0VLb/K4CFWWILZXy3
+7hsAAa7cghX1cSujqEJGEsWFZZe5SmAJ87RkTCTgC88NZm8aqIshmSM1FC6KRZ8k
+3MQavt1aRfgY32NEBI9QFtSOvoPID4i8Qsn/IOhbsPRjb+enoe+sph55L4Op331j
+oe/GgQKCAQEA2VkifgvExrquWVbiFIEfPGQUK/Rf0nUDBKbfL8DXHVfgt8jmizbA
+2anaFvEAK9rpddYFZveH0w6uu3poexcDy7ukPs0ZS0S5ae6Kge+C3GYwOJ0K18SZ
+f7PWtLYiqdU94okRt7YZAhk44151DgLc0ySgbSz4eD4OCeJaAPbIzxOpqPwAGg7v
+yAi5YkWWGuhYXN9d8M9XQ12glbpsbFWb+8jcCxooo/xoWBP6b6uuBneSG4IZgISr
+KT03kfFKgIGHiXO2Tzx4LflLaIhQWUzf/VCL5baHqXFg6nR2qTfjkmhPicnhC75j
+Q7LiZcItg5s9O0Kr9VrnDj3nCUFvY+5obQKCAQEA5Gi5AvYBI4sO//I6Y9ACZFLm
+FzWuY0K2AJCTnAuftRanV7pKThKH9y0f/kpY8v3npPks3LsC/qhzSA898R4eqE3O
+HM2tE5o61mo0SbI+rZnnmz5I7Q7wPk9XNRblPIXSe/9Q+HmzynSUFPN90PZbjZHh
+HrqnKCgJnt5eyNxVexKI7x8Z3mUKQLLFWcHlNkLi+HKxalUEhXm2DjbEVOxjVwqd
+bJU0373PkW75YVDmPlKFFANO2vMKm2C8HzEbyx755sbKOZkyuJxMt2tHVQm7depU
+ziXkxlsanGA7cijc3wRsgjqrzinmv/PtCHdpZyV2vG62aYByFfI08axttfDyww==
+-----END RSA PRIVATE KEY-----
